refactor API service
