+++ services/STARK/stark/api/dockerfile/models.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+
+from pydantic import BaseModel  # pyright: ignore[reportMissingImports]
+from typing import List
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class User(BaseModel):
+    username: str
+    groups: List[str] = []
