+++ services/STARK/stark/api/dockerfile/routers/auth.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python
+
+from fastapi import APIRouter, Depends, HTTPException, status  # pyright: ignore[reportMissingImports]
+from fastapi.security import OAuth2PasswordRequestForm  # pyright: ignore[reportMissingImports]
+
+from authentication import authenticate_user, create_access_token, get_current_user
+from models import Token, User
+
+router = APIRouter()
+
+
+@router.post("/token", response_model=Token)
+async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
+    user = authenticate_user(form_data.username, form_data.password)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token = create_access_token(
+        data={"sub": user.username, "groups": user.groups}
+    )
+    return {"access_token": access_token, "token_type": "bearer"}
+
+
+@router.get("/me")
+async def get_me(user: User = Depends(get_current_user)):
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated"
+        )
+    return {"username": user.username, "groups": user.groups}
